+++ src/gtirb_pprinter/MasmPrettyPrinter.cpp
@@ -392,7 +392,6 @@ void MasmPrettyPrinter::fixupInstruction(cs_insn& inst) {
   //   e.g  faddp st(2)  needs to be  faddp st(2),st(0)
   switch (inst.id) {
   case X86_INS_FADD:
-  case X86_INS_FADDP:
   case X86_INS_FMULP:
   case X86_INS_FDIVP:
   case X86_INS_FSUBR:
@@ -478,6 +477,13 @@ void MasmPrettyPrinter::fixupInstruction(cs_insn& inst) {
         Detail.operands[i].size = 4;
       }
     }
+  }
+
+  // Remove REPZ from REPZ RET because Masm fails with
+  // "error A2044:invalid character in file"
+  if (inst.id == X86_INS_RET && Detail.prefix[0] == X86_PREFIX_REPE) {
+    Detail.prefix[0] = 0;
+    strcpy(inst.mnemonic, "ret");
   }
 
   x86FixupInstruction(inst);